backend/requirements.txt
Django>=4.2
psycopg2-binary>=2.9
requests
PyJWT
# For DRF
djangorestframework
